split plotting file. added test with save to file
